27.1. Product Details & Smart vs Dumb Components{ The Product Card Template }
Let's really make these products come to life! Now that we have a component whose only job is to render a single product, this is going to be fun & clean.

The Product Card Template
I'll start by pasting some HTML into the template: you can copy this from the code block on this page. But there's nothing too interesting yet. We are referencing a few styles - $style['product-box'] and $style.image and we'll add a style tag soon for those. If you're wondering why I'm using the square bracket syntax, that's because JavaScript doesn't like dashes with the object property syntax: you can't say $style.product-box... which, yes, is annoying.
A little below this, we are using some product data. If you go back to the Vue dev tools and click on Catalog, each product has several fields on it, like brand, image - which is the URL to an image - name, price, and even stockQuantity.
So, for the image, we're using src="product.image" - with the : that makes this attribute dynamic - and we're rendering more data for the alt attribute, the product name and the product.price. We also have a button to view the product page... which isn't doing anything yet... then we print product.brand.
Hopefully this all feels pretty simple.
If we move over now and check the console, ooooo:

  Cannot read property product-box of undefined

Coming from product-card. Vue is telling us that the  variable is undefined... which makes sense: we don't have a style tag yet!