28.1. Loading Component{ The Important Lifecycle Hooks; mounted vs created }
Google for "vue lifecycle hooks" and click into The Vue Instance page. About half way down, you'll find a spot that talks about "Instance Lifecycle Hooks".
If you look back at catalog.vue, "lifecycle hooks" is referring to things like the mounted() function, which is called when the component is added to the page.

The Important Lifecycle Hooks
If you scroll down a bit on the docs, they have a huge diagram that you can really nerd out on: this shows all the different things that happen between a Vue instance being instantiated, mounted onto the page, rendered, updated and removed from the page if you programmatically remove a component.
Feel free to study this - it's fun stuff. But I'll highlight the three most important hooks: created, mounted and destroyed, which is called after your component is completely removed.

mounted vs created
We used mounted earlier to start our AJAX call. That means that our Vue instance was created, mounted into the DOM, and then the function was called. It turns out that a better place to load data is actually created.
Let's try this: change mounted to created and then I'll refresh to be safe. That works just fine.
The created() function is called as soon as the Vue instance for our component is instantiated. That lets us start our AJAX call as early as possible. By the time it's mounted onto the page, the products data may or may not yet be available, probably they aren't. But it doesn't really matter. And we can see this when we refresh: the products are missing for a moment.
The point is: created is the best place to do data setup like this. And mounted is the correct hook if you need to do something that will manipulate the DOM.