28.3. Loading Component{ Hiding / Showing the Loading Animation }
Hiding / Showing the Loading Animation
Ok: we only want to show the Loading component when the products AJAX call hasn't finished. The two different ways to conditionally hide or show something are v-show and v-if. In this case, especially because we're eventually going to be loading the product list multiple times when we have a search bar, let's use v-show so we can hide & show it quickly. Add v-show="". And, let's see: the easiest way to know if the products are still loading is to check if products.length === 0.
That's not a perfect solution - we'll see why later - but it's good enough for now. And when we reload... that's nice!
We can also add a v-show="" on the product-card element with products.length > 0. It's not really needed since this won't even loop if there are no products, but it balances things.
We now have dynamic products and a loading animation while the AJAX call is finishing. I'm super happy about that! But our categories are not dynamic yet. Wah, wah. Let's fix that next. But after we do, we'll explore a faster way to load them.