28.2. Loading Component{ Creating the Loading Component; Referencing an Image in CSS; Using the Component }
Creating the Loading Component
Now even though this is loading a tiny bit faster, it's still not instant. And so, to give our users confidence that our server isn't on vacation, we need a loading message.
To help us have a consistent loading message whenever we need one, let's create a shiny new Loading component. Inside components/, add a new loading.vue file. Give this a <template> with an <h4> that says: - wait for it - Loading.... That's good writing.
Let's also give this a class: :class="$style.component".
Before we add that style, create the <script> tag with export default and the options object. This only needs a name key set to Loading.
Now let's add the <style> tag with lang="scss" and module. Add just one class: .component.

Referencing an Image in CSS
If you look at the assets/ directory, it has an images/ directory with a loading.gif. In the .component class, we're going to set this as the background image. We can do that with background: url() and then the relative path to the file from here: ../../images/loading.gif. We could also add a Webpack alias for images/ if we want. Finish this with no-repeat left center and add a little padding to get the positioning just right.
Say hello to our super fancy loading component!

Using the Component
Over in index.vue, time to put it to work! Start by adding some markup to hold the loading message.
Next, import it with import Loading from '@/components/loading' and add Loading to components: the order doesn't matter.
Finally, celebrate in the template with: <Loading />.
We're not conditionally hiding and showing that yet but... there it is! Not bad!