29.1. Dynamic Categories via AJAX{ Add an Ajax Call to Sidebar }
I've gotta say, now that we have dynamic products, these hardcoded categories are starting to stress me out! So let's make those dynamic too!
If you look at /api, just like how we have a /api/products endpoint for the collection of products, we also have one for /api/categories. We can use that! And now that we know the correct way to make Ajax calls to load data, this "should" be simple! Famous last words!

Add an Ajax Call to Sidebar
Open sidebar.vue. We created a data property earlier called categories, but... it's just hardcoded. Set this to an empty array to start.
To make the AJAX call, head over to catalog.vue to celebrate one of programming's oldest arts: stealing code. Copy the entire created() function and paste it in sidebar.vue under data. We just need to change the URL to /api/categories and the data from this.products to this.categories.
Awesome! Let's... just see what happens! Refresh... whoa! I think it worked! No errors in the console... and in the dev tools, if you click on Sidebar, we have real categories data!
The data for each category is pretty simple: it has the normal @id and @type that comes from JSON-LD and it also has id and name.
Oh, and one thing I forgot to mention: when we pasted the created() function, PhpStorm automatically added the axios import for us. Sweet! Oh, but it used double quotes... which ESLint does not like. Let's fix this. Much better! If you want, you can tweak your PhpStorm settings to use single quotes automatically.