29.2. Dynamic Categories via AJAX{ v-for index and :key; Linking Properly }
v-for index and :key
Anyways, head up to the v-for directive. We learned earlier that every v-for element must have a key attribute. Until now, since our categories were hardcoded, we used the array index for the :key. But now we can be smarter because we know that each category has a unique @id property.
Let's simplify the v-for - we don't need index anymore - and then say :key="category['@id']".

Linking Properly
The last thing that doesn't work is... our links! We originally set the href to category.link... but there is no link property on the real category data.
Here's the plan: we will eventually create a separate page that will display all of the products for a specific category - the URL will be /categories/ and then the id of the category. We'll worry about making sure that page exists later, but let's get the links working now.
If you're a Symfony user, then you're used to generating a URL to a route. But from JavaScript, I'm going to keep it simple and hardcode the URL to this new page. I think it's totally ok to do this: it keeps your JavaScript simpler. The tradeoff, of course, is that if you ever changed a URL, you would need to update your JavaScript.
Ok: what we want to do here is say /category/ and then print category.id. But since we have the colon in front of href to make this dynamic, we would need to have single quotes around the string, a plus sign and then category.id.
That would work: in the browser, when I hover over categories, the right URL is display at the bottom of my browser.