29.3. Dynamic Categories via AJAX{ JavaScript Dynamic Strings }
JavaScript Dynamic Strings
But yikes! This code hurts to look at! Can we make it nicer? Of course! Replace the quotes with ticks. Now, if we need dynamic code, write ${} - so ${category.id}.
This is a superpower of modern JavaScript, not Vue. Look over now and... yes! All of the links look perfect!

Too much Dynamic Data!
This is working nicely! We have dynamic products and dynamic categories. The only thing that bothers me is all the loading! Notice that each time we refresh, I see the products and categories loading!
When we only loaded the products, that was probably okay: it was just one spot that loaded pretty fast. But also having the categories waiting to load is starting to look a bit jarring. Plus, we're eventually going to have multiple pages that will use the same categories sidebar. This means that on every page, the user will wait for the same list of categories to be fetched via AJAX. We can do better!
So next, let's investigate how we can get data from the server in a way that avoids an AJAX call. We'll use this at first to highlight the current category on the sidebar. Then later, we'll replace the AJAX call completely.