31.2.Reading Server Data & :global Classes{SASS Globals; Make our child style classes into globals}
So here's what I would love to be able to do: I want to be able to write CSS like we're doing now - but not have any classes inside .component render in a modular way. Up in the template, I want to be able to just type selected and have it work.

Make our child style classes into globals
And here's how: down in the style tag, you can tell SCSS:

  Hey! I'm using a selected class, but I don't want you to treat this like a modular class: I don't want you to add the prefix and hash.

The way you do that is by adding :global in front of it.
As soon as we do that, if I inspect element on the selected link and look at the generated CSS selector on the right... yes! It has the modular sidebar class but then it only says .selected. We're not worried about that affecting any other parts of the page because it's still inside the modular sidebar class.