31.3.Reading Server Data & :global Classes{Make our child style classes into globals}
And... we can go one step further! If you think about it, because .component will be converted into a modular class, we don't need any classes inside of it to be modular. We can move the :global up and after the .component class.
Now everything inside will not be modular. This means that the only time I should have to use the  variable is on the root element! The rest can be normal classes. This already looks much better!
Next, we have successfully referenced a global variable via a computed property. But I don't love having global variables hidden inside my code. We can do this in a better, more organized way, by refactoring it into a service!