33.2.Filtering the Products{Object Shorthand: Keys without the Key; Check it out! }
Object Shorthand: Keys without the Key
That should work... but yikes! ESLint is mad! It says:

  expected property shorthand.

This is referring to something that we've actually already done many times, but I want to highlight it in case you haven't noticed. In JavaScript, if you are trying to add a property to an object... and that property's name is the same as the variable being used for its value, you can use a shorthand syntax: just params. This sets a property named params to the params variable.

Check it out!
If we go over now, you can already see it! It reloaded! Yes! Furniture shows furniture! Breakroom shows breakroom products! Office supplies shows office supplies! And Snacks shows... uh... did this just break? Where are my snacks?
It turns out that our snacks category is currently empty! Gasp! If that's not bad enough, instead of saying - "No snacks! You're on your own! " - we see the loading screen forever.
That's... my fault. When catalog renders product-list - hold Command or Ctrl and click to jump to that - the loading is showing based on whether the products length is zero or not. An empty category looks like it's still loading!
We need to improve this: we need to truly know whether or not the AJAX call has finished! Let's make a smarter loading mechanism next!