34.1.Smarter Loading: AJAX status as State{Add loading to Catalog; Try...catch}
Oh no! The snacks category is empty! That's a huge problem on its own! To make things worse, you can't even tell! It looks like it's loading forever... while I'm sitting here getting hungrier and hungrier.
The reason is that, in the product-list component, we're showing the loading animation by checking if the products length is zero.
Using the length of an array to figure out if something is done loading doesn't really work if it's possible that the thing really can be empty in some cases! And that's totally the situation we have: sometimes a category has no products. And later when we add a search, sometimes no products will match.
So... the easy solution didn't work. What we need instead is a flag that specifically tracks whether or not the products AJAX call is finished.

Add loading to Catalog
We know Catalog is the smart component that takes care of making the AJAX request. This is means that it is also aware of whether or not we are currently making an AJAX call for the products. To track this, let's add a new data: call it loading and set it to false by default.
Now, very simply, in created(), say this.loading = true right before the AJAX call and, right after, this.loading = false.
And just like that, we have a flag that we can use to render things based on the true loading status!

Try...catch
While we're here, we can also add some simple error handling just in case the AJAX call fails. To do that, wrap all of this in a try...catch block. Then, inside catch, set this.loading = false.
If we really didn't trust our API... we could add a data called error, change that in catch to a message and render it. But with this, we will at least fail somewhat gracefully, and avoid the loader from spinning forever.
As easy as that was, this could be a bit dangerous! The problem right now is that if any of these lines have an error - like if our response doesn't have a data key on it - then the catch will be called and we will not show it. We could be hiding a bug in our code. I hate bugs! I think pizza is much better...